Rename types and modules exportation files
