+++ src/models/instance.ts
@@ -1,6 +1,6 @@
-///<reference path="../types/d2.d.ts" />
+///<reference path="../types/modules.d.ts" />
 import _ from "lodash";
-import { D2, Response } from "../types/d2-types";
+import { D2, Response } from "../types/d2";
 import { TableFilters, TableList, TablePagination } from "../types/d2-ui-components";
 import { deleteInstance, listInstances, saveNewInstance, getDataStoreData } from "./dataStore";
 import { generateUid } from "d2/uid";
