+++ src/types/modules.d.ts
@@ -1,0 +1,3 @@
+declare module "d2/uid" {
+    export function generateUid(): string;
+}
